refactor: replace dom_persor to template/parser
