+++ packages/npm-packages/runtime/src/ruby_wasm_ui/template.rb
@@ -1,0 +1,4 @@
+module RubyWasmUi
+  module Template
+  end
+end
